+++ tndb_dump.c
@@ -41,10 +41,10 @@ void timethis_end(void *tvp, const char *prefix)
 #define DUMP_DATA (1 << 0)
 
 static
-int do_walk(const char *name, unsigned flags) 
+int do_walk(const char *name, unsigned flags)
 {
     uint32_t        vlen;
-    off_t           voffs;
+    uint32_t        voffs;
     struct tndb     *db;
     struct tndb_it  it;
     char            key[TNDB_KEY_MAX + 1];
@@ -60,7 +60,7 @@ int do_walk(const char *name, unsigned flags)
         fprintf(stderr, "%s: BROKEN", name);
         return -1;
     }
-    
+
 
     if (!tndb_it_start(db, &it)) {
         fprintf(stderr, "%s: tndb_it_start failed", name);
@@ -68,7 +68,7 @@ int do_walk(const char *name, unsigned flags)
     }
 
     while ((rc = tndb_it_get_voff(&it, key, &kn, &voffs, &vlen))) {
-        
+
         if (rc < 0) {
             fprintf(stderr, "%s: tndb_it_get_voff failed", name);
             return -1;
@@ -78,17 +78,17 @@ int do_walk(const char *name, unsigned flags)
         if (flags & DUMP_DATA) {
             unsigned char *buf;
             int i;
-            
+
             buf = n_malloc(vlen + 1);
             if (tndb_read(db, voffs, buf, vlen) != (int)vlen) {
                 fprintf(stderr, "%s: %m\n", name);
                 return -1;
             }
 
-            for (i=0; i < (int)vlen; i++) 
+            for (i=0; i < (int)vlen; i++)
                 if (!isprint(buf[i]))
                     buf[i] = '.';
-            
+
             printf("DATA = %s\n\n--------------------------------------------------\n", buf);
         }
     }
@@ -98,7 +98,7 @@ int do_walk(const char *name, unsigned flags)
 
 int main(int argc, char *argv[])
 {
-    if (argc > 1) 
+    if (argc > 1)
         do_walk(argv[1], 0);
 
     return 0;
